% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_IgScan_SingleCellExperiment.R
\name{combine_IgScan_SingleCellExperiment}
\alias{combine_IgScan_SingleCellExperiment}
\title{Combine IgScan Output with Single Cell Experiment}
\usage{
combine_IgScan_SingleCellExperiment(igscan_out, sce)
}
\arguments{
\item{igscan_out}{A data frame containing the output from IgScan for a single cell dataset.}

\item{seurat_object}{A SCE object where colData will be updated with the corresponding immunogenetic information from `igscan_out`.}
}
\value{
A SCE object with updated colData, including the IgScan annotation organized by cell barcode.
}
\description{
This function integrates immunogenetic data from IgScan with a Single
Cell Experiment (SCE) object. It maps clonotype-related information and sequence data to
the colData of the SCE object based on matching barcodes.
}
\examples{
\dontrun{
updated_sce_object <- combine_IgScan_Seurat(igscan_out, sce_object)
}

}
