% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_AIRR_format.R
\name{export_AIRR_format}
\alias{export_AIRR_format}
\title{Export IgScan outputs in AIRR format}
\usage{
export_AIRR_format(object, dir, fileName = NULL)
}
\arguments{
\item{fileName}{}
}
\description{
Export IgScan outputs in AIRR format
}
