% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IgScan-package.R
\docType{package}
\name{IgScan-package}
\alias{IgScan}
\alias{IgScan-package}
\title{IgScan: A high-throughput analysis toolkit for dissecting immune repertoires in B-cell neoplasms and beyond}
\description{
IgScan is an R package for processing and analyzing high-throughput
sequencing data of B-cell immunoglobulin (Ig) repertoires, both at the
bulk and single-cell levels. The single-cell workflow supports multiple
data formats, including 10x Genomics, BD Rhapsody, and Parse Evercode.
IgScan provides detailed immunogenetic annotation, clonotype grouping,
and repertoire reconstruction, with open compatibility with AIRR
standards and other widely used tools in the field.
}
\section{Useful links}{

\itemize{
\item Package website: \url{https://nadeulab.github.io/IgScan/}
\item Report bugs at: \url{https://github.com/nadeulab/IgScan/issues}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://nadeulab.github.io/IgScan/}
  \item Report bugs at \url{https://github.com/nadeulab/IgScan/issues}
}

}
\author{
\strong{Maintainer}: Ian Márquez-López \email{iamarquez@recerca.clinic.cat}

Authors:
\itemize{
\item Ian Márquez-López
\item Ferran Nadeu \email{nadeu@recerca.clinic.cat}
}
}
\keyword{internal}
