% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_RawData.R
\name{load_RawData}
\alias{load_RawData}
\title{Load IG contigs from NGS data}
\usage{
load_RawData(sample_paths, input_format)
}
\arguments{
\item{sample_paths}{A vector of paths to input files or a vector of
directories containing the input file. Supported formats depend on the
`input_format` parameter.}

\item{input_format}{A string specifying the format of the input data, currently supporting:
'10xBCR_fasta', '10xBCR_csv', 'ParseBCR', 'BDRhapsodyBCR', 'MiXCR',
'TRUST4', 'AIRR', 'IMGT_AIRR' and 'fasta'.}
}
\value{
A list of input elements for further functions in IgScan.
  For 10xBCR_fasta and fasta formats, the raw file paths are returned as elements.
  For other sources, data frames containing the loaded and parsed data are returned.
}
\description{
This function loads raw data from various common input formats and
returns a list of elements that can be used as the 'samples_list'
parameter in other IgScan functions.

It supports data from multiple sequencing platforms and file types:
\itemize{
  \item 10x BCR = "filtered_contig_annotations.csv / filtered_contig.fa"
  \item Parse BCR = "bcr_annotation_airr.tsv"
  \item BD Rhapsody BCR = "Contigs_AIRR.tsv"
  \item MiXCR = "clonotypes.IGX.txt"
  \item TRUST4 = "barcode_report.tsv"
  \item AIRR = "airr_rearrangement.tsv"
  \item IMGT AIRR = "vquest_airr.tsv"
  \item fasta
}
}
\examples{
\dontrun{
paths <- c("path/to/bcr_annotation_airr_file1.tsv", "path/to/bcr_annotation_airr_file2.tsv")
samples_list <- load_RawData(paths, "ParseBCR")
}

}
