% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_gctree_input.R
\name{prepare_gctree_input}
\alias{prepare_gctree_input}
\title{Prepare gctree inputs from IgScan annotated datasets}
\usage{
prepare_gctree_input(
  object,
  outDir,
  mode = "single_cell",
  cloneID = "C1",
  chain = "HL"
)
}
\arguments{
\item{object}{A data.frame, \code{SingleCellExperiment}, or \code{Seurat} object with IgScan annotation.}

\item{outDir}{A string. Path to the directory where output files should be written.}

\item{mode}{A string. Either \code{"single_cell"} or \code{"bulk"} depending on the dataset type. Default is \code{"single_cell"}.}

\item{cloneID}{A string. The clonotype identifier to extract and process. Note that the \code{cloneID} parameter must
match a value in the clonotype identifier field in the IgScan objects.}

\item{chain}{A string. One of \code{"H"} (heavy chain), \code{"L"} (light chain), or \code{"HL"} (paired chains). Default is \code{"HL"}.
Only needed when \code{mode} is set to \code{"single_cell"}.}
}
\value{
The function returns no value in R but writes the necessary input files for \code{gctree} into the \code{outDir}.
}
\description{
This function takes as input an IgScan annotated dataframe coming from bulk
NGS or single cell workflows, or a single cell object with IgScan annotation,
and prepares the input files for running the \code{gctree} program for BCR phylogenetic
inference. Once the input files have been generated, the user can run \code{gctree}
following the instructions in XXXX jupyter notebook provided with the IgScan package.
}
\details{
The function generates the following files in the specified \code{outDir}:
\itemize{
\item \code{abundances.csv}: Abundance of each subclone including the germline (GL).
\item \code{idmap.txt}: Mapping of subclone IDs to individual barcodes (or rownames).
\item \code{subclones.fasta}: FASTA file of the aligned somatic and germline VDJ sequences.
\item \code{deduplicated.phylip}: Alignment file in PHYLIP format for input to \code{gctree}.
}
}
\examples{
\dontrun{
# Example with a SingleCellExperiment object
prepare_gctree_input(sce_object, outDir = "results/clone1/", mode = "single_cell", cloneID = "C1", chain = "HL")

# Example with a Seurat object
prepare_gctree_input(seurat_object, outDir = "results/clone2/", mode = "single_cell", cloneID = "C5", chain = "H")

# Example with a bulk IgScan-annotated dataframe
prepare_gctree_input(bulk_df, outDir = "results/clone3/", mode = "bulk", cloneID = "CloneA")
}

}
