% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalculate_IDs_single_cell.R
\name{recalculate_IDs_single_cell}
\alias{recalculate_IDs_single_cell}
\title{Recalculate IgScan IDs for Single Cell Data}
\usage{
recalculate_IDs_single_cell(single_cell_object)
}
\arguments{
\item{single_cell_object}{A single-cell object combined with IgScan annotation
in either `SingleCellExperiment` or `Seurat` format.}
}
\value{
A single cell object with the IgScan clonotype-related IDs updated in the metadata.
}
\description{
This function recalculates and standardizes IgScan clonotype-related IDs
in single-cell sequencing datasets. Using this function may be of interest
when a IgScan-annotated single cell object has been filtered and some cells
have been removed. It supports input objects from both the `SingleCellExperiment`
and `Seurat` classes.
}
\examples{
\dontrun{
recalculated_object <- recalculate_IDs_single_cell(sc_object)
}

}
