% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescue_single_chain_cells.R
\name{rescue_single_chain_cells}
\alias{rescue_single_chain_cells}
\title{Rescue cells with incomplete BCR (single chains)}
\usage{
rescue_single_chain_cells(
  single_cell_object,
  group_col = "orig.ident",
  threads = 1
)
}
\arguments{
\item{single_cell_object}{A single-cell object combined with IgScan annotation
in either \code{SingleCellExperiment} or \code{Seurat} format. This function also supports
IgScan data.frames coming from the analysis of single cell data.}

\item{group_col}{A vector with the name of the column or columns containing the grouping variable/s.
Default is 'orig.ident', thus rescuing and recalculating BCR IDs based in each sample independently.}

\item{threads}{The number of threads to perform single_chain cells rescuing. Default is 1.}
}
\value{
The input single cell object (or IgScan data.frame) with updated
metadata, where the \code{completeBCR}field in some "single chain" cells may now be marked as "Yes_rescue".
}
\description{
This function identifies and rescues cells with incomplete BCR information (labeled
as "Single_chain_1" or "Single_chain_2") by attempting to infer and assign clonotype
and subclonotype IDs based on related cells that have complete BCR information. It
supports both \code{SingleCellExperiment} and \code{Seurat} objects. After rescuing single-chain
entries, the function also recalculates clonotype-related identifiers.
}
\examples{
\dontrun{
sce <- rescue_single_chain_cells(sce, group_col = "sample", threads = 4)
seurat <- rescue_single_chain_cells(seurat, group_col = "orig.ident", threads = 2)
igscan_df <- rescue_single_chain_cells(igscan_df, group_col = "SampleID", threads = 2)
}

}
