% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_IgBlast_from_RawData.R
\name{Run_IgBlast_from_RawData}
\alias{Run_IgBlast_from_RawData}
\title{Run IgBlast on raw NGS data}
\usage{
Run_IgBlast_from_RawData(
  sample_paths,
  sample_labels,
  input_format,
  data_type = "single_cell",
  Evalue_cutoff = NULL,
  annotate_C = T,
  threads_IgBlast = 1,
  outputDir = NULL,
  run_IgBlast_report = F,
  threads = 1
)
}
\arguments{
\item{sample_paths}{A vector of paths to input files or a vector of
directories containing the input file. Supported formats depend on the
`input_format` parameter.}

\item{sample_labels}{A vector of sample labels corresponding to `sample_paths`.
Must have the same length as `sample_paths`.}

\item{input_format}{A string specifying the format of the input data, currently supporting:
'10xBCR_fasta', '10xBCR_csv', 'ParseBCR', 'BDRhapsodyBCR', 'MiXCR',
'TRUST4', 'AIRR', 'IMGT_AIRR' and 'fasta'.}

\item{data_type}{A string indicating the type of data. Options: 'single_cell' and 'bulk'. Default is 'single_cell'.}

\item{Evalue_cutoff}{A numeric value specifying the E-value cutoff for IgBLAST alignment.
Default is `NULL` (no cutoff applied).}

\item{annotate_C}{Logical. If `TRUE`, the function includes the annotation of
constant regions using the NCBI IG C region database.}

\item{threads_IgBlast}{An integer specifying the number of threads for running IgBLAST.
Default is 1.}

\item{outputDir}{The directory to store intermediate and output files.
Defaults to the working directory if not provided.}

\item{run_IgBlast_report}{Logical. If `TRUE`, generates a PDF report summarizing
IgBLAST results.}

\item{threads}{An integer specifying the number of threads for processing samples.
Default is 1.}
}
\value{
A list containing the paths of all generated IgBLAST output files. If
  `run_IgBlast_report = TRUE`, also includes the path to the PDF report.
}
\description{
This function processes raw sequencing data and runs IgBLAST to perform
immunoglobulin reannotation using the IMGT databases. The function supports
multiple input formats and generates annotated outputs, including reports
if wanted by the user.
}
\details{
This function performs the following steps:

   1. Validates the input data and format.

   2. Converts input data to FASTA format if needed.

   3. Runs IgBLAST for V(D)J-(C) reannotation and alignment.

   4. Generates an IgBlast summary report if specified.
}
\note{
Ensure that IgBLAST and its required reference databases are installed and
  properly configured before running this function.
}
\examples{
\dontrun{
# Example 1: Running IgBLAST with FASTA inputs
Run_IgBlast_from_RawData(
  sample_paths = list("path/to/sample1.fasta", "path/to/sample2.fasta"),
  sample_labels = c("Sample1", "Sample2"),
  input_format = "fasta",
  data_type = "single_cell",
  outputDir = "path/to/outDir")

# Example 2: Running IgBLAST with 10x Genomics CSV inputs and generating a IgBlast report
Run_IgBlast_from_RawData(
  sample_paths = list("path/to/sample1_filtered_contig_annotations.csv"),
  sample_labels = c("Sample1"),
  input_format = "10xbcr_csv",
  data_type = "single_cell",
  outputDir = "path/to/outDir",
  run_IgBlast_report = TRUE)
}

}
