% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_IgScan_FullWorkflow.R
\name{Run_IgScan_FullWorkflow}
\alias{Run_IgScan_FullWorkflow}
\title{Run IgScan Full Workflow}
\usage{
Run_IgScan_FullWorkflow(
  sample_paths,
  sample_labels,
  Evalue_cutoff = NULL,
  annotate_C = TRUE,
  threads = 1,
  threads_IgBlast = 1,
  case_labels = NULL,
  input_format,
  analysis_mode = "single",
  material_type = "rna",
  v_primer = "full_length",
  data_type = "single_cell",
  min_reads = 2,
  remove_tmp = TRUE,
  outputDir = NULL,
  hc_similarity_cutoff = 0.2,
  hc_mode = "average",
  cdr3_mode = "nt",
  cdr3_InDel_correction_mode = "soft_filter",
  annotate_CLL_immGen = FALSE,
  annotate_satellite_subsets = TRUE,
  annotate_ags = FALSE,
  rescue_single_chain = FALSE,
  relaxed_rescue = FALSE
)
}
\arguments{
\item{sample_paths}{A vector of paths to input files or a vector of
directories containing the input file. Supported formats depend on the
\code{input_format} parameter.}

\item{sample_labels}{A vector of sample labels corresponding to \code{sample_paths}.
Must have the same length as \code{sample_paths}.}

\item{Evalue_cutoff}{A numeric value specifying the E-value cutoff for IgBLAST alignment.
Default is \code{NULL} (no cutoff applied).}

\item{annotate_C}{Logical. If \code{TRUE}, the function performs an IgBlast annotation of
constant regions using the NCBI IG C region database.}

\item{threads}{An integer specifying the number of threads for processing samples.
Default is 1.}

\item{threads_IgBlast}{An integer specifying the number of threads for running IgBLAST.
Default is 1.}

\item{case_labels}{A vector of case labels matching the indexes of \code{sample_labels}.
Required when analysis mode is set to 'joint'. Default is NULL.}

\item{input_format}{A string specifying the format of the input data, currently supporting:
'10xBCR_fasta', '10xBCR_csv', 'ParseBCR', 'BDRhapsodyBCR', 'MiXCR',
'TRUST4', 'AIRR', 'IMGT_AIRR' and 'fasta'.}

\item{analysis_mode}{Defines the mode of analysis to be performed. Options are
"single" and "joint". In "single" mode, each sample is annotated independently.
In "joint" mode, samples are grouped by case based on the "case_list" vector
for a case-level joint annotation. Default is 'single'.}

\item{material_type}{The biological source of material. Options: 'dna' and 'rna'. Note that for
matrial_type='rna', unproductive sequences are not expected, and will be directly removed. Default is 'rna'.}

\item{v_primer}{The primer sequence used for the V-region amplification.
Options: 'full_length', 'fr1', 'fr2' and 'fr3'. Only required for \code{data_type='bulk'}. Default is
'full_length'. Note that sequences with unexpected length pattern based on the chosen primer
will be directly excluded from the analysis.}

\item{data_type}{The type of data. Options: 'single_cell', 'bulk' or 'missionbio'. Default is 'single_cell'.}

\item{min_reads}{The minimum number of reads required to keep a sequence in the downstream analyses.
Only needed for bulk NGS data. Default is 2.}

\item{remove_tmp}{Logical value indicating whether to remove the temporary files (all files
except the annotation_results directory). Default is TRUE.}

\item{outputDir}{Path to the directory containing the previous IgScan outputs coming from the
Run_IgBlast_from_RawData function. There is NO default value for this parameter.}

\item{hc_similarity_cutoff}{Hierarchical clustering distance in CDR3 identity to consider two sequences
clonally related. Default is 0.2, meaning that up to 20 percent of mismatch is accepted.}

\item{hc_mode}{The aglomeration method to be used in the hierarchical clustering. Options are: 'single',
'complete' and 'average' (UPGMA). Default is 'average'.}

\item{cdr3_mode}{The type of CDR3 sequence to be used for the hierarchical clustering.
Options are 'nt' (nucleotides) and 'aa' (amino acids). Default is 'nt'.}

\item{cdr3_InDel_correction_mode}{The degree of stringency in correcting InDels in the CDR3 region. Options
are 'no' (no CDR3 InDel correction, recommended for highly polyclonal samples), 'hard_filter' (high stringency
but computationally slower) and 'soft_filter' (less stringent but much faster). Default is 'soft_filter'.}

\item{annotate_CLL_immGen}{Logical value indicating whether to annotate CLL-related immunogenetic
information such as CLL stereotyped subsets and the R110 mutation. Default is FALSE.}

\item{annotate_satellite_subsets}{Logical value indicating whether to annotate Satellite CLL stereotyped subsets.
Only needed if \code{annotate_CLL_immGen} is set to TRUE. Default is TRUE.}

\item{annotate_ags}{Logical value indicating whether to annotate IGH Acquired N-Glycosylation Sites (AGS).
Default is FALSE.}

\item{rescue_single_chain}{Logical value indicating whether to hard assign cells with one chain detected to the most
plausible complete clonotype (see \code{rescue_single_chain_cells} function documentation for more details). Default is FALSE.}

\item{relaxed_rescue}{Logical value indicating whether to apply a relaxed clonotype rescue mode.
Only needed if \code{rescue_single_chain} is set to \code{TRUE}.The default (recommended) is \code{FALSE}, requiring exact
V(D)J nucleotide sequence matches to rescue single-chain cells. If set to \code{TRUE}, which we recommend only
for the analysis of Mission Bio single-cell V(D)J data, rescue is performed at the clonotype level rather
than by exact nucleotide sequence identity.}
}
\value{
A list containing an IgScan annotation dataframe for every sample analyzed, which are also
saved in the output directory. These dataframes can be directly passed to further IgScan functions.
}
\description{
This function provides a streamlined pipeline to process raw
sequencing data, perform IgBLAST reannotation, and run IgScan to generate
advanced immunogenetic analyses. It combines the functionalities of
\code{Run_IgBlast_from_RawData} and \code{Run_IgScan_Annotation} into a single step,
allowing users to process raw scRNA-seq/bulk NGS data, annotate immunoglobulin
sequences, and analyze clonotypes in a unified workflow. In addition, the workflow is
also compatible with single-cell V(D)J data generated using the Mission Bio platform,
provided that the IgScan input has been produced using our dedicated raw-data
preprocessing pipeline (see IgScan GitHub).

It supports data from multiple sequencing platforms and file types:
\itemize{
\item 10x BCR = "filtered_contig_annotations.csv / filtered_contig.fa"
\item Parse BCR = "bcr_annotation_airr.tsv"
\item BD Rhapsody BCR = "Contigs_AIRR.tsv"
\item MiXCR = "clonotypes.IGX.txt"
\item TRUST4 = "barcode_report.tsv"
\item AIRR = "airr_rearrangement.tsv"
\item IMGT AIRR = "vquest_airr.tsv"
\item fasta
}
}
\details{
This function performs the following steps:
\enumerate{
\item Validates the input data and format.
\item Converts input data to FASTA format if needed.
\item Runs IgBLAST for V(D)J-(C) reannotation and alignment.
\item Processes IgBLAST output with the IgScan workflow to perform an advanced
immunogenetic annotation of the data, cluster sequences into clonotypes and
formats the output based on the type of experiment (bulk or single cell).
}
}
\examples{
\dontrun{
# Example: Run the complete pipeline on FASTA inputs
results <- Run_IgScan_FullWorkflow(
  sample_paths = list("path/to/sample1.fasta", "path/to/sample2.fasta"),
  sample_labels = c("Sample1", "Sample2"),
  analysis_mode = "single",
  Evalue_cutoff = 100,
  annotate_C = T,
  threads_IgBlast = 2,
  input_format = "fasta",
  material_type = "dna",
  v_primer = "full_length",
  data_type = "bulk",
  min_reads = 2,
  remove_tmp = TRUE,
  outputDir = "path/output/dir/",
  hc_similarity_cutoff = 0.2,
  hc_mode = "average",
  cdr3_mode = "nt",
  cdr3_InDel_correction_mode = "soft_filter",
  annotate_CLL_immGen = TRUE,
  annotate_satellite_subsets = FALSE,
  threads = 4)
}

}
