% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_IgScan_Seurat.R
\name{combine_IgScan_Seurat}
\alias{combine_IgScan_Seurat}
\title{Combine IgScan Output with Seurat Object}
\usage{
combine_IgScan_Seurat(igscan_out, seurat_object)
}
\arguments{
\item{igscan_out}{A data frame containing the output from IgScan for a single cell dataset.}

\item{seurat_object}{A Seurat object where metadata will be updated with the corresponding immunogenetic information from `igscan_out`.}
}
\value{
A Seurat object with updated metadata, including the IgScan annotation organized by cell barcode.
}
\description{
This function integrates immunogenetic data from IgScan with a Seurat
object. It maps clonotype-related information and sequence data to the metadata of
the Seurat object based on matching barcodes.
}
\examples{
\dontrun{
updated_seurat_object <- combine_IgScan_Seurat(igscan_out, seurat_object)
}

}
