% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_BCR_contam_bulk.R
\name{filter_BCR_contam_bulk}
\alias{filter_BCR_contam_bulk}
\title{Filter BCR contamination in bulk IG-NGS data}
\usage{
filter_BCR_contam_bulk(
  igscan_data_frame,
  batch_col = NULL,
  case_col = NULL,
  contamination_cutoff = 10,
  contamination_clone_cutoff = 80,
  remove_contamination = F,
  recalc_column = "SampleID",
  threads = 1
)
}
\arguments{
\item{igscan_data_frame}{An IgScan output dataframe coming from the annotation
of bulk NGS data.}

\item{batch_col}{A string with the name of the column identifying the \code{batch} of processing
and/or sequencing for each sample. Each \code{batch} will be evaluated independently. Default is NULL,
meaning that all sequences in the data frame will be considered from the same batch.}

\item{case_col}{A string with the name of the column identifying the \code{case} or \code{individual} for
each sample. Default is NULL, meaning that each sample will be considered from different individuals.}

\item{contamination_cutoff}{A numeric value specifying the contamination ratio threshold.
Default is 10, higher values result in stricter contamination detection.}

\item{contamination_clone_cutoff}{A numeric value defining what percentage of contaminated sequences within
a clonotype is allowed (i.e., sequences sharing the same clonotype ID). If the proportion of contaminated
sequences exceeds this threshold, all sequences beloning to the clonptype will be flagged as \code{OUT_CLONOTYPE}.
Default is 80 percent.}

\item{remove_contamination}{Logical. If \code{TRUE}, sequences flagged as contaminated are
removed from the returned data frame (default is \code{FALSE}).}

\item{recalc_column}{A string/vector of strings with the name of the column/s to use for ID
recalculation. Needed when remove_contamination=TRUE. Default is \code{SampleID}.}

\item{threads}{The number of threads to perform BCR ID recalculation. Default is 1.}
}
\value{
The input \code{igscan_data_frame} including three new columns:
\itemize{
\item Contamination_FLAG: Indicates whether a cell passed contamination checks (\code{PASS}), was flagged as contaminated (\code{OUT}),
or was involved in crossed contamination (\code{OUT_X})
\item Contamination_Freq: The frequency of the contaminated sequence in the original sample (dominant).
\item Contamination_Sample: The dominant sample responsible for contamination.
}

If \code{remove_contamination = TRUE}, sequences flagged as \code{OUT} or \code{OUT_X} are removed from
the data frame and immunogenetic IDs are recalculated with \code{recalculate_IDs_bulk()}.
}
\description{
This function identifies, flags, and optionally removes potential contamination
between samples processed and/or sequenced together in a IgScan-annotated bulk
NGS dataset. Contamination is assessed by analyzing the frequency of identical
IG sequences identified in different samples A contamination cutoff can be specified,
indicating that if a clonotype is observed X times more in one sample than another,
it will be flagged as contamination in the latter sample.
}
\examples{
\dontrun{
cont_filtered_igscan <- filter_BCR_contam_bulk(igscan_data_frame = igscan_df, bath_col = "batchID", case_col = "case", contamination_cutoff = 30, threads = 4)
}

}
