% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_BCR_contam_bulk.R
\name{filter_BCR_contam_bulk}
\alias{filter_BCR_contam_bulk}
\title{Filter BCR contamination in bulk NGS data}
\usage{
filter_BCR_contam_bulk(
  igscan_data_frame,
  batch_col,
  contamination_cutoff = 10,
  remove_contamination = F
)
}
\arguments{
\item{igscan_data_frame}{An IgScan output dataframe coming from the annotation
of bulk NGS data.}

\item{batch_col}{A string with the name of the column identifying the `batch` of processing
and/or sequencing for each sample. Each `batch` will be evaluated independently.}

\item{contamination_cutoff}{A numeric value specifying the contamination ratio threshold.
Default is 10, higher values result in stricter contamination detection.}

\item{remove_contamination}{Logical. If `TRUE`, cells flagged as contaminated are
removed from the returned object (default is `FALSE`).}
}
\value{
The input `igscan_data_frame` including three new columns:
\itemize{
  \item Contamination_FLAG: Indicates whether a cell passed contamination checks (`PASS`), was flagged as contaminated (`OUT`),
  or was involved in crossed contamination (`OUT_X`)
  \item Contamination_Ratio: The ratio of clonal overlap between the contaminated and dominant samples.
  \item Contamination_Sample: The dominant sample responsible for contamination.
}

  If `remove_contamination = TRUE`, cells flagged as `OUT` or `OUT_X` are removed from
  the data frame and immunogenetic IDs are recalculated with `recalculate_IDs_bulk()`.
}
\description{
This function identifies, flags, and optionally removes potential contamination
between samples processed and/or sequenced together in a IgScan-annotated bulk
NGS dataset. Contamination is assessed by analyzing the frequency of identical
IG sequences identified in different samples A contamination cutoff can be specified,
indicating that if a clonotype is observed X times more in one sample than another,
it will be flagged as contamination in the latter sample.
}
\examples{
\dontrun{
# Example with a SingleCellExperiment object:
cont_filtered_igscan <- filter_BCR_contam_bulk(igscan_data_frame = igscan_df,, bath_col = "batchID",contamination_cutoff = 30)
}

}
