% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_BCR_contam_single_cell.R
\name{filter_BCR_contam_single_cell}
\alias{filter_BCR_contam_single_cell}
\title{Filter BCR-based contamination in single-cell data}
\usage{
filter_BCR_contam_single_cell(
  single_cell_list,
  sample_col = "orig.ident",
  batch_col = NULL,
  case_col = NULL,
  contamination_cutoff = 10,
  contamination_clone_cutoff = 80,
  remove_contamination = F,
  recalc_column = "orig.ident",
  threads = 1
)
}
\arguments{
\item{single_cell_list}{A list of `SingleCellExperiment` or `Seurat` objects containing IgScan
annotation for single cell data.}

\item{sample_col}{A string with the name of the column identifying `sample`.
Default is 'orig.ident'.}

\item{batch_col}{A string with the name of the column identifying the `batch` of processing
and/or sequencing for each sample. Each `batch` will be evaluated independently. Default is NULL,
meaning that all samples in the single_cell_object list will be considered from the same batch.}

\item{case_col}{A string with the name of the column identifying the `case` or `individual` for
each sample. Default is NULL, meaning that each sample will be considered from different individuals.}

\item{contamination_cutoff}{A numeric value specifying the contamination ratio threshold.
Default is 10, higher values result in stricter contamination detection.}

\item{contamination_clone_cutoff}{A numeric value specifying which percentage of contamination
is allowed within a clonotype to ........ .}

\item{remove_contamination}{Logical. If `TRUE`, cells flagged as contaminated are
removed from the returned object (default is `FALSE`).}

\item{recalc_column}{A string/vector of strings with the name of the column/s to use for ID
recalculation. Needed when remove_contamination=TRUE. Default is `orig.ident`.}

\item{threads}{The number of threads to perform BCR ID recalculation. Default is 1.}
}
\value{
The input `single_cell_list` with each object including three new columns:
\itemize{
  \item Contamination_FLAG: Indicates whether a cell passed contamination checks (`PASS`), was flagged as contaminated (`OUT`),
  or was involved in crossed contamination (`OUT_X`)
  \item Contamination_Ratio: The ratio of clonal overlap between the contaminated and dominant samples.
  \item Contamination_Sample: The dominant sample responsible for contamination.
}

  If `remove_contamination = TRUE`, cells flagged as `OUT` or `OUT_X` are removed from
  the objects and immunogenetic IDs are recalculated with `recalculate_IDs_single_cell()`.
}
\description{
This function identifies, flags, and optionally removes potential contamination
between samples processed and/or sequenced together fors cRNA-seq+scBCR-seq.
Contamination is assessed by analyzing the distribution of identicalBCR clonotypes
between samples. A contamination cutoff can be specified, indicating that if a
clonotype is observed X times more in one sample tag than another, it will be
flagged as contamination in the latter sample. The function works with both
`SingleCellExperiment` and `Seurat` objects.
}
\examples{
\dontrun{
# Example with a SingleCellExperiment object:
cont_filtered_sce <- filter_BCR_contam_single_cell(single_cell_list = sc_list, bath_col = "batchID", case_col = "case", contamination_cutoff = 30, threads = 4)
}

}
