% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalculate_IDs_bulk.R
\name{recalculate_IDs_bulk}
\alias{recalculate_IDs_bulk}
\title{Recalculate IgScan IDs for Bulk NGS Data}
\usage{
recalculate_IDs_bulk(igscan_data_frame)
}
\arguments{
\item{igscan_data_frame}{An IgScan output dataframe coming from the annotation
of bulk NGS data.}
}
\value{
An IgScan output data frame with updated clonotype-related IDs and
cumulative number of reads and frequencies.
}
\description{
This function recalculates and standardizes clonotype-related IDs
in IgScan outputs of bulk NGS datasets. Using this function may be of interest
when a IgScan-annotated dataframe has been filtered and some sequences
have been removed. It also updates the cumulative number of reads and frequencies
of the different IDs.
}
\examples{
\dontrun{
updated_igscan_out <- recalculate_IDs_bulk(igscan_output_df)
}
}
